+++ changelog.md
@@ -8,6 +8,7 @@
 
 -   Add `href` property to `Button`, `HeaderItem`, `ListItem` and `MenuItem`
     to make they render as anchor element
+-   [TooltipHandler] Add `openDuration` property
 
 ## [8.0.0-alpha.6] - 2019-05-04